refactor: #123 - auth 도메인 기준으로 canonicalizeEmail 위치 이동
