+++ src/features/auth/utils/canonicalizeEmail.test.ts
