+++ src/features/auth/utils/canonicalizeEmail.ts
